+++ ErsatzTV.Core/Scheduling/PlayoutBuilder.cs
@@ -91,6 +91,11 @@ namespace ErsatzTV.Core.Scheduling
                                 await _mediaCollectionRepository.GetMultiCollectionItems(
                                     collectionKey.MultiCollectionId ?? 0);
                             return Tuple(collectionKey, multiCollectionItems);
+                        case ProgramScheduleItemCollectionType.SmartCollection:
+                            List<MediaItem> smartCollectionItems =
+                                await _mediaCollectionRepository.GetSmartCollectionItems(
+                                    collectionKey.SmartCollectionId ?? 0);
+                            return Tuple(collectionKey, smartCollectionItems);
                         default:
                             return Tuple(collectionKey, new List<MediaItem>());
                     }
@@ -514,6 +519,7 @@ namespace ErsatzTV.Core.Scheduling
                         CollectionType = collectionKey.CollectionType,
                         CollectionId = collectionKey.CollectionId,
                         MultiCollectionId = collectionKey.MultiCollectionId,
+                        SmartCollectionId = collectionKey.SmartCollectionId,
                         MediaItemId = collectionKey.MediaItemId,
                         EnumeratorState = maybeEnumeratorState[collectionKey]
                     });
@@ -535,6 +541,7 @@ namespace ErsatzTV.Core.Scheduling
                      && a.CollectionType == collectionKey.CollectionType
                      && a.CollectionId == collectionKey.CollectionId
                      && a.MultiCollectionId == collectionKey.MultiCollectionId
+                     && a.SmartCollectionId == collectionKey.SmartCollectionId
                      && a.MediaItemId == collectionKey.MediaItemId);
 
             CollectionEnumeratorState state = maybeAnchor.Match(
@@ -661,6 +668,11 @@ namespace ErsatzTV.Core.Scheduling
                     CollectionType = item.CollectionType,
                     MultiCollectionId = item.MultiCollectionId
                 },
+                ProgramScheduleItemCollectionType.SmartCollection => new CollectionKey
+                {
+                    CollectionType = item.CollectionType,
+                    SmartCollectionId = item.SmartCollectionId
+                },
                 _ => throw new ArgumentOutOfRangeException(nameof(item))
             };
 
@@ -669,6 +681,7 @@ namespace ErsatzTV.Core.Scheduling
             public ProgramScheduleItemCollectionType CollectionType { get; set; }
             public int? CollectionId { get; set; }
             public int? MultiCollectionId { get; set; }
+            public int? SmartCollectionId { get; set; }
             public int? MediaItemId { get; set; }
         }
     }