+++ ErsatzTV.Infrastructure/Data/TvContext.cs
@@ -59,6 +59,7 @@ namespace ErsatzTV.Infrastructure.Data
         public DbSet<Collection> Collections { get; set; }
         public DbSet<CollectionItem> CollectionItems { get; set; }
         public DbSet<MultiCollection> MultiCollections { get; set; }
+        public DbSet<SmartCollection> SmartCollections { get; set; }
         public DbSet<ProgramSchedule> ProgramSchedules { get; set; }
         public DbSet<ProgramScheduleItem> ProgramScheduleItems { get; set; }
         public DbSet<Playout> Playouts { get; set; }